fix: Stabilize vite dev-mode ITs
Wait for an element on the main page to load before
commencing tests.
